feat: code refactor and add test case
